pyburst_excesspower: correct undersampling rate
The undersampling rate has to be set to so that the new Nyquist
frequency lies above the total frequency content spanned from zero.
Undersampling is, in effect, shifting a bandlimited frequency series from
its lowest frequency down to DC. The previous setting would have cut
into the BL frequency content because it was using the FWHM bandwidth
rather than the full bandwidth -- this corrects this factor of two.